chore: move test utils under internal/utils/test
