+++ internal/util/test/controller_manager.go
@@ -1,4 +1,4 @@
-package util
+package test
 
 import (
 	"context"
