+++ internal/util/test/crds.go
@@ -1,4 +1,4 @@
-package util
+package test
 
 import (
 	"bytes"
